+++ apps/next/src/popup/index.tsx
@@ -16,7 +16,7 @@ import { Children, lazy, ReactElement, Suspense } from 'react';
 import { I18nextProvider } from 'react-i18next';
 import { HashRouter as Router } from 'react-router-dom';
 import { initI18n, i18next } from '@core/common';
-import React from 'react';
+import { Providers } from './providers';
 
 // Initialize translations
 initI18n();
@@ -28,32 +28,6 @@ const App = lazy(() => {
 });
 
 const root = createRoot(document.getElementById('popup') as HTMLElement);
-
-export const Providers = ({
-  providers,
-  children,
-}: {
-  providers: ReactElement[];
-  children: ReactElement;
-}) => {
-  const renderProvider = (
-    renderedProviders: ReactElement[],
-    renderedChildren: ReactElement,
-  ) => {
-    const [provider, ...restProviders] = renderedProviders;
-
-    if (provider) {
-      return React.cloneElement(
-        provider,
-        undefined,
-        renderProvider(restProviders, renderedChildren),
-      );
-    }
-
-    return renderedChildren;
-  };
-  return renderProvider(providers, children);
-};
 
 browser.tabs.query({ active: true }).then(() => {
   root.render(
